refactor: name of ui to common
